Size is not grandeur, and territory does not make a nation

+++ 0x00-hello_world/6-size.c
@@ -13,7 +13,7 @@ int main(void)
 	float floatT;
 	long int longIntT;
 	char charT;
-	unsigned int longLongIntT;
+	long long int longLongIntT;
 
 	printf("Size of a char: %zu bytes\n", sizeof(charT));
 	printf("Size of an int: %zu bytes\n", sizeof(intT));
